requirements.txt
openai ~= 1.2
python-dotenv ~= 1.0
requests ~= 2.31
streamlit ~= 1.27

# Linters and formatters (use the latest versions)
ruff
